Remove duplicate writeError.
Both the livereload server and the watcher were listening to the `watcher`'s `error`
event which caused the same error to be printed twice.

The `didError` in `livereload-server` is superfluous (it cannot run
without `wacther` running and `watcher` already has a `didError` that
does the right thing).

Moved the tests from livereload-server-tests to watcher-tests.
