+++ lib/tasks/server/livereload-server.js
@@ -41,7 +41,6 @@ module.exports = Task.extend({
 
     // Reload on file changes
     this.watcher.on('change', this.didChange.bind(this));
-    this.watcher.on('error',  this.didError.bind(this));
 
     // Start LiveReload server
     return this.listen(options.liveReloadPort)
@@ -77,13 +76,5 @@ module.exports = Task.extend({
         message: 'live-reload'
       });
     }
-  },
-
-  didError: function(error) {
-    this.ui.writeError(error);
-    this.analytics.trackError({
-      description: error.message + ' ' + error.stack,
-      isFatal:     false
-    });
   }
 });
