+++ tests/unit/models/watcher-test.js
@@ -8,6 +8,7 @@ var MockWatcher  = require('../../helpers/mock-watcher');
 var Watcher = require('../../../lib/models/watcher');
 var EOL = require('os').EOL;
 var chalk = require('chalk');
+var BuildError = require('../../helpers/build-error');
 
 describe('Watcher', function() {
   var ui;
@@ -75,17 +76,66 @@ describe('Watcher', function() {
   });
 
   describe('watcher:error', function() {
-    beforeEach(function() {
+    it('tracks errors', function() {
       watcher.emit('error', {
         message: 'foo',
         stack: new Error().stack
       });
-    });
 
-    it('tracks errors', function() {
       assert.deepEqual(analytics.trackErrors, [{
         description: 'foo'
       }]);
+    });
+
+    it('emits without error.file', function() {
+      subject.didError(new BuildError({
+        file: 'someFile',
+        message: 'buildFailed'
+      }));
+
+      var outs = ui.output.split(EOL);
+
+      assert.equal(outs[0], chalk.red('File: someFile'));
+      assert.equal(outs[1], chalk.red('buildFailed'));
+    });
+
+    it('emits with error.file with error.line without err.col', function() {
+      subject.didError(new BuildError({
+        file: 'someFile',
+        line: 24,
+        message: 'buildFailed'
+      }));
+
+      var outs = ui.output.split(EOL);
+
+      assert.equal(outs[0], chalk.red('File: someFile (24)'));
+      assert.equal(outs[1], chalk.red('buildFailed'));
+    });
+
+    it('emits with error.file without error.line with err.col', function() {
+      subject.didError(new BuildError({
+        file: 'someFile',
+        col: 80,
+        message: 'buildFailed'
+      }));
+      var outs = ui.output.split(EOL);
+
+      assert.equal(outs[0], chalk.red('File: someFile'));
+      assert.equal(outs[1], chalk.red('buildFailed'));
+    });
+
+    it('emits with error.file with error.line with err.col', function() {
+      subject.didError(new BuildError({
+        file: 'someFile',
+        line: 24,
+        col: 80,
+        message: 'buildFailed'
+      }));
+
+      var outs = ui.output.split(EOL);
+
+      assert.equal(outs[0], chalk.red('File: someFile (24:80)'));
+      assert.equal(outs[1], chalk.red('buildFailed'));
     });
   });
